deploy: setup ability to deploy from locally built lambda
Cross compiling on user's machine. Then mounting the volume into the
docker container and simply executing a `cp` command in order to place
the binary into the expected place `/asset-output/bootstrap`.

This only requires an `alpine` image which is small and a custom
Dockerfile isn't needed for this approach.

Added the ability to choose between that approach and building inside
the container. The alternate way still needs to be implemented.