feat: cargo chef in a working state with modifications
Outlined in the readme and code cargo chef is not fully working as
expected and we work around this by rebuilding on container start.
However it is much faster than without cargo chef already and for very
fast turn around cycle it is recommended to use the option to cross
compile on the user's machine instead of inside a container.